+++ src/components/PrimaryButton/index.tsx
@@ -9,9 +9,8 @@ const PrimaryButton: React.FC<PrimaryButtonProps> = ({ children, className, onCl
     <button
       {...rest}
       onClick={onClick}
-      className={`px-4 py-2 text-sm font-medium leading-5 text-center text-white transition-colors duration-150 bg-primary border border-transparent rounded-lg
-           hover:bg-opacity-80 focus:outline-none focus:ring active:bg-primary'
-      } ${className}`}>
+      className={`px-4 py-2 text-sm font-medium leading-5 text-center text-white rounded-lg
+      } transition-colors duration-150 bg-primary border border-transparent hover:bg-opacity-80 focus:outline-none focus:ring active:bg-primary ${className}`}>
       {children}
     </button>
   );
